src/test/java/com/poxete/biblioteca_pessoal/usecase/SalvarAutorUseCaseTest.java
package com.poxete.biblioteca_pessoal.usecase;

import com.poxete.biblioteca_pessoal.ConfiguraBaseEmMemoria;
import com.poxete.biblioteca_pessoal.service.AutorService;
import com.poxete.biblioteca_pessoal.service.dto.AutorDTO;
import org.junit.jupiter.api.BeforeAll;
import org.junit.jupiter.api.Test;
import org.junit.jupiter.api.TestInstance;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.boot.test.context.SpringBootTest;

import static com.google.common.truth.Truth.assertThat;

@SpringBootTest(properties = "spring.profiles.active=test")
@TestInstance(TestInstance.Lifecycle.PER_CLASS)
class SalvarAutorUseCaseTest {
    @Autowired
    ConfiguraBaseEmMemoria configuraBaseEmMemoria;

    @Autowired
    SalvarAutorUseCase salvarAutorUseCase;

    @Autowired
    AutorService autorService;

    @BeforeAll
    public void setup() {
        configuraBaseEmMemoria.configurarBancoDeDados();
    }

    @Test
    void deveSalvarAutor() {
        var autor = salvarAutorUseCase.salvarAutor(new AutorDTO("Poxete2", "pxt", false));

        assertThat(autor.getNome()).isEqualTo("Poxete2");
        assertThat(autor.getOutrosNomes().size()).isEqualTo(1);
    }
}
